refactored to use laravel sanctum stateful cookies

+++ database/migrations/2024_05_08_010202_create_email_queues_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("template");
             $table->string('email');
-            $table->integer('send_after_ts');
+            $table->unsignedBigInteger('send_after_ts');
             $table->json('data');
             $table->timestamps();
         });
